- Updating list options for inventory levels - Fixing exports - Updating descriptions for inventory levels
