+++ models/index.ts
@@ -11,8 +11,10 @@ export * from "./discount";
 export * from "./discount_code";
 export * from "./fulfillment";
 export * from "./gift_card";
+export * from "./inventory_level";
 export * from "./line_item";
 export * from "./line_item_property";
+export * from "./location";
 export * from "./order";
 export * from "./payment_details";
 export * from "./price_rule";
